Created pedalSensor Structures
I added the sensor structures for reading the pedals. They are exactly the same as the old code. I think a lot of this stuff can be kept pretty similar, except for the TV stuff which will most likely go in the place of the old PedalMap file. I can handle implementing that once Ben is ready, but for now I will keep porting over the rest of the stuff that TV should not really impact.